--- examples/demos/bugfix/index.ts
@@ -1,10 +1,0 @@
-export { MapRender as color } from './color';
-export { MapRender as data_shake } from './data-shake';
-export { MapRender as event_legend } from './event_legend';
-export { MapRender as mutiPolygon } from './muti-polygon';
-export { MapRender as remove_muti_layer } from './remove-muti-layer';
-export { MapRender as setData } from './set-data';
-export { MapRender as size } from './size';
-export { MapRender as text } from './text_offsets';
-export { MapRender as tile_update } from './tile_update';
-export { MapRender as touchend } from './touchend';
