--- examples/demos/tile/index.ts
@@ -1,6 +1,0 @@
-export { MapRender as image } from './normal';
-export { MapRender as perf } from './perf';
-export { MapRender as rgb } from './rgb';
-export { MapRender as usa } from './usa';
-export { MapRender as vector } from './vector';
-export { MapRender as worldline } from './world';
