+++ examples/demos_next/components/index.ts
@@ -1,0 +1,5 @@
+export { layerPopup } from './layer-popup';
+export { marker } from './marker';
+export { popup } from './popup';
+export { swipe } from './swipe';
+export { zoom } from './zoom';
