+++ examples/demos_next/tile/index.ts
@@ -1,0 +1,6 @@
+export { rasterTile } from './raster-tile';
+export { terrainRGB } from './terrain-rgb';
+export { vectorCity } from './vector-city';
+export { vectorLine } from './vector-line';
+export { vectorPerf } from './vector-perf';
+export { vectorUSA } from './vector-usa';
