+++ examples/demos_next/tile/vector-line.ts
@@ -1,20 +1,18 @@
 import { LineLayer, Scene, Source } from '@antv/l7';
-import * as allMap from '@antv/l7-maps';
 import { Protocol } from 'pmtiles';
-import type { RenderDemoOptions } from '../../types';
+import type { TestCase } from '../../types';
+import { CaseScene } from '../../utils';
 
 const protocol = new Protocol();
 Scene.addProtocol('pmtiles', protocol.tile);
 
-export function MapRender(options: RenderDemoOptions) {
-  const scene = new Scene({
-    id: 'map',
-    renderer: options.renderer,
-    map: new allMap[options.map]({
-      style: 'light',
+export const vectorLine: TestCase = async (options) => {
+  const scene = await CaseScene({
+    ...options,
+    mapConfig: {
       center: [116.420818, 39.923852],
       zoom: 14,
-    }),
+    },
   });
 
   const source = new Source(
@@ -37,11 +35,7 @@ export function MapRender(options: RenderDemoOptions) {
     opacity: 0.8,
   });
 
-  scene.on('loaded', () => {
-    scene.addLayer(linelayer);
+  scene.addLayer(linelayer);
 
-    if (window['screenshot']) {
-      window['screenshot']();
-    }
-  });
-}
+  return scene;
+};
